refactor(test): 이벤트 테스트 패키지 변경
* 이벤트 테스트들을 이벤트 패키지로 옮김
